+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Config changelog
 
+## [2.1.1](http://github.com/ivanoff/2valid/tree/2.0.2) (2016-08-25)
+[Full Changelog](http://github.com/ivanoff/2valid/compare/2.0.2...2.1.1)
+
+**What Was Done:**
+
+- add validate with no creating model
+- add notRequired field to result
+
 ## [2.0.2](http://github.com/ivanoff/2valid/tree/2.0.2) (2016-08-24)
 [Full Changelog](http://github.com/ivanoff/2valid/compare/2.0.1...2.0.2)